requirements.txt
tensorflow==2.14.0
streamlit==1.31.1
numpy
matplotlib
Pillow